scholarship_platform_backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Automat==25.4.16
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
cssselect==1.3.0
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
google-generativeai==0.8.3
packaging==25.0
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
queuelib==1.8.0
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.3
service-identity==24.2.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tldextract==5.3.0
tqdm==4.67.1
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
w3lib==2.3.1
Werkzeug==3.1.3
zope.interface==7.2
